Extract common parts from Parameter Store and Secrets Manager integrations (#251)
