+++ v3/spring-cloud-aws-v3-autoconfigure/src/main/java/io/awspring/cloud/v3/autoconfigure/config/parameterstore/ParameterStoreConfigDataLoader.java
@@ -14,16 +14,12 @@
  * limitations under the License.
  */
 
-package io.awspring.cloud.v3.autoconfigure.parameterstore;
+package io.awspring.cloud.v3.autoconfigure.config.parameterstore;
 
-import java.lang.reflect.Field;
-import java.lang.reflect.Modifier;
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.List;
 
-import io.awspring.cloud.v3.paramstore.ParameterStorePropertySource;
-import org.apache.commons.logging.Log;
+import io.awspring.cloud.v3.autoconfigure.config.BootstrapLoggingHelper;
+import io.awspring.cloud.v3.parameterstore.ParameterStorePropertySource;
 import software.amazon.awssdk.services.ssm.SsmClient;
 
 import org.springframework.boot.context.config.ConfigData;
@@ -31,8 +27,6 @@ import org.springframework.boot.context.config.ConfigDataLoader;
 import org.springframework.boot.context.config.ConfigDataLoaderContext;
 import org.springframework.boot.context.config.ConfigDataResourceNotFoundException;
 import org.springframework.boot.logging.DeferredLogFactory;
-import org.springframework.util.ClassUtils;
-import org.springframework.util.ReflectionUtils;
 
 /**
  * {@link ConfigDataLoader} for AWS Parameter Store.
@@ -43,11 +37,10 @@ import org.springframework.util.ReflectionUtils;
  */
 public class ParameterStoreConfigDataLoader implements ConfigDataLoader<ParameterStoreConfigDataResource> {
 
-	private final DeferredLogFactory logFactory;
-
 	public ParameterStoreConfigDataLoader(DeferredLogFactory logFactory) {
-		this.logFactory = logFactory;
-		reconfigureLoggers(logFactory);
+		BootstrapLoggingHelper.reconfigureLoggers(logFactory,
+				"io.awspring.cloud.v3.parameterstore.ParameterStorePropertySource",
+				"io.awspring.cloud.v3.autoconfigure.config.parameterstore.ParameterStorePropertySources");
 	}
 
 	@Override
@@ -66,31 +59,6 @@ public class ParameterStoreConfigDataLoader implements ConfigDataLoader<Paramete
 		catch (Exception e) {
 			throw new ConfigDataResourceNotFoundException(resource, e);
 		}
-	}
-
-	private void reconfigureLoggers(DeferredLogFactory logFactory) {
-		// loggers in these classes must be static non-final
-		List<Class<?>> loggers = new ArrayList<>();
-		loggers.add(ParameterStorePropertySources.class);
-		// class may be not present if parameterstore module is not on the classpath
-		if (ClassUtils.isPresent("io.awspring.cloud.v3.paramstore.ParameterStorePropertySource", null)) {
-			loggers.add(ParameterStorePropertySource.class);
-		}
-
-		loggers.forEach(it -> reconfigureLogger(it, logFactory));
-	}
-
-	static void reconfigureLogger(Class<?> type, DeferredLogFactory logFactory) {
-		ReflectionUtils.doWithFields(type, field -> {
-
-			field.setAccessible(true);
-			field.set(null, logFactory.getLog(type));
-
-		}, ParameterStoreConfigDataLoader::isUpdateableLogField);
-	}
-
-	private static boolean isUpdateableLogField(Field field) {
-		return !Modifier.isFinal(field.getModifiers()) && field.getType().isAssignableFrom(Log.class);
 	}
 
 }
