+++ v3/spring-cloud-aws-v3-autoconfigure/src/main/java/io/awspring/cloud/v3/autoconfigure/config/parameterstore/ParameterStorePropertySources.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package io.awspring.cloud.v3.autoconfigure.parameterstore;
+package io.awspring.cloud.v3.autoconfigure.config.parameterstore;
 
-import io.awspring.cloud.v3.paramstore.ParameterStorePropertySource;
+import io.awspring.cloud.v3.parameterstore.ParameterStorePropertySource;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import software.amazon.awssdk.services.ssm.SsmClient;
