+++ v3/spring-cloud-aws-v3-autoconfigure/src/main/java/io/awspring/cloud/v3/autoconfigure/config/secretsmanager/SecretsManagerConfigDataLoader.java
@@ -14,16 +14,12 @@
  * limitations under the License.
  */
 
-package io.awspring.cloud.v3.autoconfigure.secretsmanager;
+package io.awspring.cloud.v3.autoconfigure.config.secretsmanager;
 
-import java.lang.reflect.Field;
-import java.lang.reflect.Modifier;
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.List;
 
+import io.awspring.cloud.v3.autoconfigure.config.BootstrapLoggingHelper;
 import io.awspring.cloud.v3.secretsmanager.SecretsManagerPropertySource;
-import org.apache.commons.logging.Log;
 import software.amazon.awssdk.services.secretsmanager.SecretsManagerClient;
 
 import org.springframework.boot.context.config.ConfigData;
@@ -31,8 +27,6 @@ import org.springframework.boot.context.config.ConfigDataLoader;
 import org.springframework.boot.context.config.ConfigDataLoaderContext;
 import org.springframework.boot.context.config.ConfigDataResourceNotFoundException;
 import org.springframework.boot.logging.DeferredLogFactory;
-import org.springframework.util.ClassUtils;
-import org.springframework.util.ReflectionUtils;
 
 /**
  * Loads config data from AWS Secret Manager.
@@ -44,11 +38,10 @@ import org.springframework.util.ReflectionUtils;
  */
 public class SecretsManagerConfigDataLoader implements ConfigDataLoader<SecretsManagerConfigDataResource> {
 
-	private final DeferredLogFactory logFactory;
-
 	public SecretsManagerConfigDataLoader(DeferredLogFactory logFactory) {
-		this.logFactory = logFactory;
-		reconfigureLoggers(logFactory);
+		BootstrapLoggingHelper.reconfigureLoggers(logFactory,
+				"io.awspring.cloud.v3.secretsmanager.SecretsManagerPropertySource",
+				"io.awspring.cloud.v3.autoconfigure.config.secretsmanager.SecretsManagerPropertySources");
 	}
 
 	@Override
@@ -67,31 +60,6 @@ public class SecretsManagerConfigDataLoader implements ConfigDataLoader<SecretsM
 		catch (Exception e) {
 			throw new ConfigDataResourceNotFoundException(resource, e);
 		}
-	}
-
-	private void reconfigureLoggers(DeferredLogFactory logFactory) {
-		// loggers in these classes must be static non-final
-		List<Class<?>> loggers = new ArrayList<>();
-		loggers.add(SecretsManagerPropertySources.class);
-		// class may be not present if secrets manager module is not on the classpath
-		if (ClassUtils.isPresent("io.awspring.cloud.v3.secretsmanager.SecretsManagerPropertySource", null)) {
-			loggers.add(SecretsManagerPropertySource.class);
-		}
-
-		loggers.forEach(it -> reconfigureLogger(it, logFactory));
-	}
-
-	static void reconfigureLogger(Class<?> type, DeferredLogFactory logFactory) {
-		ReflectionUtils.doWithFields(type, field -> {
-
-			field.setAccessible(true);
-			field.set(null, logFactory.getLog(type));
-
-		}, SecretsManagerConfigDataLoader::isUpdateableLogField);
-	}
-
-	private static boolean isUpdateableLogField(Field field) {
-		return !Modifier.isFinal(field.getModifiers()) && field.getType().isAssignableFrom(Log.class);
 	}
 
 }
