+++ v3/spring-cloud-aws-v3-autoconfigure/src/main/java/io/awspring/cloud/v3/autoconfigure/config/secretsmanager/SecretsManagerPropertySources.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.awspring.cloud.v3.autoconfigure.secretsmanager;
+package io.awspring.cloud.v3.autoconfigure.config.secretsmanager;
 
 import io.awspring.cloud.v3.secretsmanager.SecretsManagerPropertySource;
 import org.apache.commons.logging.Log;
