+++ v3/spring-cloud-aws-v3-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -5,15 +5,15 @@ io.awspring.cloud.v3.autoconfigure.ses.SesAutoConfiguration
 
 # ConfigData Location Resolvers
 org.springframework.boot.context.config.ConfigDataLocationResolver=\
-io.awspring.cloud.v3.autoconfigure.parameterstore.ParameterStoreConfigDataLocationResolver,\
-io.awspring.cloud.v3.autoconfigure.secretsmanager.SecretsManagerConfigDataLocationResolver
+io.awspring.cloud.v3.autoconfigure.config.parameterstore.ParameterStoreConfigDataLocationResolver,\
+io.awspring.cloud.v3.autoconfigure.config.secretsmanager.SecretsManagerConfigDataLocationResolver
 
 # ConfigData Loaders
 org.springframework.boot.context.config.ConfigDataLoader=\
-io.awspring.cloud.v3.autoconfigure.parameterstore.ParameterStoreConfigDataLoader,\
-io.awspring.cloud.v3.autoconfigure.secretsmanager.SecretsManagerConfigDataLoader
+io.awspring.cloud.v3.autoconfigure.config.parameterstore.ParameterStoreConfigDataLoader,\
+io.awspring.cloud.v3.autoconfigure.config.secretsmanager.SecretsManagerConfigDataLoader
 
 # Failure Analyzers
 org.springframework.boot.diagnostics.FailureAnalyzer=\
-io.awspring.cloud.v3.autoconfigure.parameterstore.ParameterStoreMissingKeysFailureAnalyzer, \
-io.awspring.cloud.v3.autoconfigure.secretsmanager.SecretsManagerMissingKeysFailureAnalyzer
+io.awspring.cloud.v3.autoconfigure.config.parameterstore.ParameterStoreMissingKeysFailureAnalyzer, \
+io.awspring.cloud.v3.autoconfigure.config.secretsmanager.SecretsManagerMissingKeysFailureAnalyzer
