+++ v3/spring-cloud-aws-v3-parameter-store/src/test/java/io/awspring/cloud/v3/parameterstore/ParameterStorePropertySourceTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.awspring.cloud.v3.paramstore;
+package io.awspring.cloud.v3.parameterstore;
 
 import org.junit.jupiter.api.Test;
 import software.amazon.awssdk.services.ssm.SsmClient;
